+++ src/main/java/ru/clevertec/Main.java
@@ -2,9 +2,12 @@ package ru.clevertec;
 
 import ru.clevertec.animal.data.AnimalDto;
 import ru.clevertec.animal.entity.Animal;
-import ru.clevertec.animal.service.BaseService;
+import ru.clevertec.animal.proxy.cache.IBaseCache;
+import ru.clevertec.animal.proxy.cache.imp.LRUCache;
+import ru.clevertec.animal.service.IBaseService;
 import ru.clevertec.animal.service.impl.AnimalServiceImpl;
 
+import java.lang.reflect.Field;
 import java.util.UUID;
 
 public class Main {
@@ -19,14 +22,24 @@ public class Main {
 //        System.out.println(mapper.merge(animal, animalDto).toString());
 //        System.out.println(ObjectValidator.validate(animal));
 //        System.out.println(ObjectValidator.validate(animalDto));
-        BaseService<AnimalDto> service = new AnimalServiceImpl();
-
+        IBaseService<AnimalDto> service = new AnimalServiceImpl();
+//        service.get(UUID.fromString("657e186a-1a01-48a9-a839-116b81b85057"));
 //
-//        System.out.println(service.get(UUID.fromString("657e186a-1a01-48a9-a839-116b81b85057")));
-//        service.create(animalDto);
-//        animalDto.setName("Страус");
-//        service.update(UUID.fromString("b357dd88-5ff7-4485-bc7e-da81cca3fa6e"), animalDto);
-//        service.delete(UUID.fromString("b357dd88-5ff7-4485-bc7e-da81cca3fa6e"));
-        System.out.println(service.getAll());
+        System.out.println(service.get(UUID.fromString("657e186a-1a01-48a9-a839-116b81b85057")));
+        service.getAll().stream()
+                .forEach(System.out::println);
+        service.create(animalDto);
+        service.getAll().stream()
+                .forEach(System.out::println);
+        animalDto.setName("Страус");
+        service.getAll().stream()
+                .forEach(System.out::println);
+        service.update(UUID.fromString("318f55d9-607c-40d2-930a-2329b4dc5304"), animalDto);
+        service.getAll().stream()
+                .forEach(System.out::println);
+        service.delete(UUID.fromString("318f55d9-607c-40d2-930a-2329b4dc5304"));
+        service.getAll().stream()
+                .forEach(System.out::println);
     }
+
 }
