+++ src/main/java/ru/clevertec/animal/entity/Animal.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldNameConstants;
-import ru.clevertec.animal.validator.NumericFields;
-import ru.clevertec.animal.validator.TextFields;
 
 import java.util.UUID;
