+++ src/main/java/ru/clevertec/animal/service/IBaseService.java
@@ -3,7 +3,7 @@ package ru.clevertec.animal.service;
 import java.util.List;
 import java.util.UUID;
 
-public interface BaseService<T> {
+public interface IBaseService<T> {
 
     T get(UUID uuid);
     List<T> getAll();
