+++ src/main/java/ru/clevertec/animal/validator/annotation/TextFields.java
@@ -1,4 +1,4 @@
-package ru.clevertec.animal.validator;
+package ru.clevertec.animal.validator.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,7 +7,8 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface NumericFields {
-    double min() default 0.0000000001;
-    double max() default Double.POSITIVE_INFINITY;
+public @interface TextFields {
+    int minLength() default 0;
+
+    int maxLength() default 30;
 }
